Remove dependency on glob from ristretto/z pkg
glog pollutes the global FlagSet in the stdlib flag package. This causes
collisions that result in panics.

glog starts a goroutine during init() that runs forever which wastes
resources.

glog makes syscalls during the init() phase that block the startup
of the program. This causes delays from Windows programs that start
as services.